+++ server/src/controllers/organizations.controller.ts
@@ -1,15 +1,4 @@
-import {
-  BadRequestException,
-  Body,
-  Controller,
-  Get,
-  NotFoundException,
-  Param,
-  Patch,
-  Post,
-  UseGuards,
-  Query,
-} from '@nestjs/common';
+import { Body, Controller, Get, NotFoundException, Param, Patch, Post, UseGuards, Query } from '@nestjs/common';
 import { OrganizationsService } from '@services/organizations.service';
 import { AppConfigService } from '@services/app_config.service';
 import { decamelizeKeys } from 'humps';
@@ -22,7 +11,7 @@ import { PoliciesGuard } from 'src/modules/casl/policies.guard';
 import { User as UserEntity } from 'src/entities/user.entity';
 import { ConfigService } from '@nestjs/config';
 import { MultiOrganizationGuard } from 'src/modules/auth/multi-organization.guard';
-import { OrganizationCreateDto } from '@dto/organization-create.dto';
+import { OrganizationCreateDto, OrganizationUpdateDto } from '@dto/organization';
 
 @Controller('organizations')
 export class OrganizationsController {
@@ -123,12 +112,9 @@ export class OrganizationsController {
   @UseGuards(JwtAuthGuard, PoliciesGuard)
   @CheckPolicies((ability: AppAbility) => ability.can('updateOrganizations', UserEntity))
   @Patch()
-  async update(@Body() body, @User() user) {
-    if (body?.name?.length > 25) {
-      throw new BadRequestException('name cannot be longer than 25 characters');
-    }
-    await this.organizationsService.updateOrganization(user.organizationId, body);
-    return {};
+  async update(@Body() organizationUpdateDto: OrganizationUpdateDto, @User() user) {
+    await this.organizationsService.updateOrganization(user.organizationId, organizationUpdateDto);
+    return;
   }
 
   @UseGuards(JwtAuthGuard, PoliciesGuard)
